P4 3392 volumetric price change part 1 (#386)
* P4-3392 bumping 3rd libraries prior to main changes.

* P4-3392 initial steps moving towards using BigDecimal instead of double for price adjustments. More to do at time of commit.

* P4-3392 next set of steps towards using BigDecimal instead of double for price adjustments. DB migration left to do.

* P4-3392 migration to alter column type for price adjustments to support larger number of decimal points.

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/domain/price/PriceAdjustment.kt
@@ -1,5 +1,6 @@
 package uk.gov.justice.digital.hmpps.pecs.jpc.domain.price
 
+import java.math.BigDecimal
 import java.time.LocalDateTime
 import java.util.UUID
 import javax.persistence.Column
@@ -44,5 +45,5 @@ data class PriceAdjustment(
   val effectiveYear: Int,
 
   @Column(name = "multiplier", nullable = false)
-  val multiplier: Double
+  val multiplier: BigDecimal
 )
