+++ src/test/kotlin/uk/gov/justice/digital/hmpps/pecs/jpc/domain/price/PriceAdjustmentRepositoryTest.kt
@@ -20,7 +20,7 @@ internal class PriceAdjustmentRepositoryTest {
   fun `can create price adjustment for Serco and retrieve by supplier`() {
     assertThat(repository.findBySupplier(Supplier.SERCO)).isNull()
 
-    val persisted = repository.saveAndFlush(PriceAdjustment(supplier = Supplier.SERCO, multiplier = 1.5, effectiveYear = 2020))
+    val persisted = repository.saveAndFlush(PriceAdjustment(supplier = Supplier.SERCO, multiplier = 1.5.toBigDecimal(), effectiveYear = 2020))
 
     assertThat(repository.findBySupplier(Supplier.SERCO)).isEqualTo(persisted)
   }
@@ -29,7 +29,7 @@ internal class PriceAdjustmentRepositoryTest {
   fun `price adjustment is in progress for Serco and exists by supplier`() {
     assertThat(repository.existsPriceAdjustmentBySupplier(Supplier.SERCO)).isFalse
 
-    repository.saveAndFlush(PriceAdjustment(supplier = Supplier.SERCO, multiplier = 1.2, effectiveYear = 2021))
+    repository.saveAndFlush(PriceAdjustment(supplier = Supplier.SERCO, multiplier = 1.2.toBigDecimal(), effectiveYear = 2021))
 
     assertThat(repository.existsPriceAdjustmentBySupplier(Supplier.SERCO)).isTrue
   }
@@ -38,7 +38,7 @@ internal class PriceAdjustmentRepositoryTest {
   fun `can create price adjustment for GEOAmey and retrieve by supplier`() {
     assertThat(repository.findBySupplier(Supplier.GEOAMEY)).isNull()
 
-    val persisted = repository.saveAndFlush(PriceAdjustment(supplier = Supplier.GEOAMEY, multiplier = 2.0, effectiveYear = 2022))
+    val persisted = repository.saveAndFlush(PriceAdjustment(supplier = Supplier.GEOAMEY, multiplier = 2.0.toBigDecimal(), effectiveYear = 2022))
 
     assertThat(repository.findBySupplier(Supplier.GEOAMEY)).isEqualTo(persisted)
   }
@@ -47,7 +47,7 @@ internal class PriceAdjustmentRepositoryTest {
   fun `can only have one price adjustment for a supplier at any given point in time`() {
     assertThat(repository.findBySupplier(Supplier.SERCO)).isNull()
 
-    val priceAdjustment = repository.saveAndFlush(PriceAdjustment(supplier = Supplier.SERCO, multiplier = 1.5, effectiveYear = 2020))
+    val priceAdjustment = repository.saveAndFlush(PriceAdjustment(supplier = Supplier.SERCO, multiplier = 1.5.toBigDecimal(), effectiveYear = 2020))
 
     assertThatThrownBy { repository.saveAndFlush(priceAdjustment.copy(id = UUID.randomUUID())) }
       .isInstanceOf(DataIntegrityViolationException::class.java)
